+++ inst/htmlwidgets/echarts4r.js
@@ -8,7 +8,7 @@ HTMLWidgets.widget({
     
     var initialized = false;
 
-    var chart;
+    var chart,opts;
     
     return {
 
@@ -38,7 +38,11 @@ HTMLWidgets.widget({
         }
         
         chart = echarts.init(document.getElementById(el.id), x.theme, {renderer: x.renderer});
-        chart.setOption(x.opts);
+        
+        opts = x.opts;
+        
+        if(x.draw === true)
+          chart.setOption(opts);
         
         
         // shiny callbacks
@@ -162,6 +166,10 @@ HTMLWidgets.widget({
       getChart: function(){
         return chart;
       },
+      
+      getOpts: function(){
+        return opts;
+      },
 
       resize: function(width, height) {
 
@@ -188,7 +196,30 @@ function get_e_charts(id){
   return(echarts);
 }
 
+function get_e_charts_opts(id){
+
+  var htmlWidgetsObj = HTMLWidgets.find("#" + id);
+
+  var echarts;
+
+  if (typeof htmlWidgetsObj != 'undefined') {
+    echarts = htmlWidgetsObj.getOpts();
+  }
+
+  return(echarts);
+}
+
 if (HTMLWidgets.shinyMode) {
+  
+  // DRAW
+  Shiny.addCustomMessageHandler('e_draw_p',
+    function(data) {
+      var chart = get_e_charts(data.id);
+      var opts = get_e_charts_opts(data.id);
+      if (typeof chart != 'undefined') {
+        chart.setOption(opts);
+      }
+  });
   
   // HIGHLIGHT AND DOWNPLAY